requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
easyocr>=1.7.0
ultralytics>=8.0.0
numpy
pandas
mysql-connector-python>=8.0.0
PyYAML
streamlit
firebase-admin>=6.2.0
pandas>=2.0.0
